tools/fpga_simulator/requirements.txt
vpython
pyserial
